+++ test/production/app-dir/graceful-degrade/app/browser-crash-error-boundary/page.tsx
