+++ test/production/app-dir/graceful-degrade/app/browser-crash/page.tsx
