+++ test/production/app-dir/graceful-degrade/app/layout.tsx
